+++ test/filter.spec.js
@@ -1,0 +1,8 @@
+import 'babel-polyfill';
+import createUniqueValuesTest from './filtering/createUniqueValues';
+import filterTest from './filtering/filter';
+
+describe('filtering', () => {
+  describe('createUniqueValues', createUniqueValuesTest);
+  describe('filter', filterTest);
+});